completed internet security assignment whole

+++ internet-security/rsa/rsa.c
@@ -2,9 +2,9 @@
 #include<stdlib.h>
 #include<time.h>
 #include<math.h>
-int prime(int n)
+long long prime(long long n)
 {
-        int i, flag = 0;
+        long long i, flag = 0;
         for (i = 2; i <= n / 2; ++i) {
                 if (n % i == 0) {
                         flag = 1;
@@ -13,10 +13,10 @@ int prime(int n)
         }
         return flag;
 }
-int find_prime(int x)
+long long find_prime(long long x)
 {
-        printf("inside prime: %d \n",x);
-        int i;
+        printf("inside prime: %lld \n",x);
+        long long i;
         for (i = x+1; i; i++) {
                 if (prime(i)==0)
                         return i;
@@ -24,22 +24,22 @@ int find_prime(int x)
         }
         return 0;
 }
-int positive_modulo(int i, int n)
+long long positive_modulo(long long i, long long n)
 {
         return (i % n + n) % n;
 }
-int mode(int n, int p, int r)
+long long mode(long long n, long long p, long long r)
 {
-        int i;
-        int a1 = positive_modulo(n, r);
-        int q = 1;
+        long long i;
+        long long a1 = positive_modulo(n, r);
+        long long q = 1;
         for (i = 1; i <= p; i++) {
                 q *= a1;
                 q = positive_modulo(q, r);
         }
         return q;
 }
-int gcd(int a, int b)
+long long gcd(long long a, long long b)
 {
         if (b == 0)
                 return a;
@@ -47,26 +47,26 @@ int gcd(int a, int b)
                 return gcd(b, positive_modulo(a, b));
 }
 
-int coprime(int e, int n)
+long long coprime(long long e, long long n)
 {
         return (gcd(e, n) == 1);
 }
 struct p_text
 {
         char char_s[20];
-        int value;
+        long long value;
 };
-int main()
+long long main()
 {
         //encryption
         FILE *fp;
         time_t t;
-        int k, b_size, i, j, pro = 0, max = 0, n1, n2, n, e, p, d, powe, m;
+        long long k, b_size, i, j, pro = 0, max = 0, n1, n2, n, e, p, d, powe, m;
         char ch;
         struct p_text b_no[100], cip[100], dec[100];
         fp = fopen("plain.txt", "r");
         printf("Enter block size:");
-        scanf("%d", &b_size);
+        scanf("%lld", &b_size);
         i = 0;
         j = 0;
         while ((ch = fgetc(fp)) != EOF) {
@@ -94,12 +94,12 @@ int main()
                 printf("%s ", b_no[i].char_s);
         printf("\n");
         for (i = 0; i <= j; i++)
-                printf("%d ", b_no[i].value);
+                printf("%lld ", b_no[i].value);
         printf("\n");
         fclose(fp);
         for (i = 0; i < b_size; i++)
                 max = max + pow(26, i) * 25;
-        printf("Maximum value can be: %d", max);
+        printf("Maximum value can be: %lld", max);
         //random number generation
         srand((unsigned) time(&t));
         printf("\n");
@@ -107,28 +107,28 @@ int main()
         n1 = find_prime(sqrt(max));
         n2 = find_prime(n1);
 
-        printf("two prime numbers:%d %d\n", n1, n2);
+        printf("two prime numbers:%lld %lld\n", n1, n2);
         n = n1 * n2;
-        printf("Value of n: %d\n", (n1 * n2));
-        printf("Value of phi: %d\n", ((n1 - 1) * (n2 - 1)));
+        printf("Value of n: %lld\n", (n1 * n2));
+        printf("Value of phi: %lld\n", ((n1 - 1) * (n2 - 1)));
         p = ((n1 - 1) * (n2 - 1));
         e = find_prime(sqrt(p));
         d = find_prime(e);
 
         for (i = 0; i < p; i++) {
                 if ((i * e) % p == 1) {
-                        printf("Value of D: %d", i);
+                        printf("Value of D: %lld", i);
                         break;
                 }
         }
         d = i;
         printf("\n");
-        printf("Public key: {%d ,%d}\n", e, n);
-        printf("Private key: {%d ,%d}\n", d, n);
+        printf("Public key: {%lld ,%lld}\n", e, n);
+        printf("Private key: {%lld ,%lld}\n", d, n);
         printf("ciphertext :");
         for (i = 0; i <= j; i++) {
                 cip[i].value = mode(b_no[i].value, e, n);
-                printf("%d ", cip[i].value);
+                printf("%lld ", cip[i].value);
         }
 
         //Decryption
@@ -136,7 +136,7 @@ int main()
         printf("Decrypted: ");
         for (i = 0; i <= j; i++) {
                 dec[i].value = mode(cip[i].value, d, n);
-                printf("%d ", dec[i].value);
+                printf("%lld ", dec[i].value);
         }
         for (i = 0; i <= j; i++) {
                 k = 0;
